moved hearing-details-header.component, hearingservice  to shared.
