+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/self-test-journey/pages/check-your-computer/check-your-computer.component.html
@@ -2,9 +2,9 @@
   <h1 class="govuk-heading-xl" i18n="@@check_your_computer_h1">
     Check your computer
   </h1>
-<!--   <div>
+  <div>
     <app-hearing-details-header></app-hearing-details-header>
-  </div> -->
+  </div>
 </div>
 <div *ngIf="isIndividual">
   <h1 class="govuk-heading-xl" i18n="@@check_your_computer_h1">
