+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/shared/hearing-details-header/hearing-details-header.component.css
