+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/shared/hearing-details-header/hearing-details-header.component.spec.ts
@@ -1,20 +1,20 @@
 import { JourneyBase } from 'src/app/modules/base-journey/journey-base';
 import { HearingDetailsHeaderComponent } from './hearing-details-header.component';
 import { TestBed, fakeAsync, tick } from '@angular/core/testing';
-import { LongDatetimePipe } from '../../shared/date-time.pipe';
+import { LongDatetimePipe } from '../date-time.pipe';
 import { Hearing } from '../../base-journey/participant-suitability.model';
 import { HearingDetails } from '../../representative-journey/services/hearing-details';
-import { HearingService } from '../../representative-journey/services/hearing.service';
+import { HearingService } from '../services/hearing.service';
 import { MutableRepresentativeSuitabilityModel } from '../../representative-journey/mutable-representative-suitability.model';
 import { RepresentativeJourney } from '../../representative-journey/representative-journey';
-import { LongDatePipe } from '../../shared/date.pipe';
+import { LongDatePipe } from '../date.pipe';
 
 describe('HearingDetailsHeaderComponent', () => {
   const hearing = new HearingDetails('id', 'case name', 'case number', 'case type', new Date(2029, 4, 3, 12, 30));
   let hearingService: jasmine.SpyObj<HearingService>;
   let component: HearingDetailsHeaderComponent;
 
-  beforeAll(() => {
+  beforeEach(() => {
     hearingService = jasmine.createSpyObj<HearingService>(['getHearing']);
     hearingService.getHearing.and.returnValue(Promise.resolve(hearing));
